separation des fonctions de convertion du scanf dans des fichiers separees
